user login e register, compra e venda de acoes

+++ invest_monitor_api/src/main/java/com/investmonitor/api/invest_monitor_api/dto/ValueWithRaw.java
@@ -1,3 +1,5 @@
 package com.investmonitor.api.invest_monitor_api.dto;
 
-public record ValueWithRaw(String raw, double value) {}
+import java.math.BigDecimal;
+
+public record ValueWithRaw(String raw, BigDecimal value) {}
